Move spark session fixture to root conftest.py to be usable outside spark subpackage.
